Add status line(mainly for showing the line number)
vim "set nu" can be used to show the line number but has
an issue which is, line numbers will be copied as well when
you copy contents using mouse.

So given what we want is just showing line numbers, here
we found a plugin show a cool status line which contains
line numbers.

Signed-off-by: Mark Zhang <ace119@163.com>